% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_events.R
\name{get_issue_events}
\alias{get_issue_events}
\title{Get GitHub issue events}
\usage{
get_issue_events(issue_url, limit = Inf)
}
\arguments{
\item{issue_url}{GitHub issue URL}

\item{limit}{Number of events to return. Defaults to \code{Inf}}
}
\description{
Get GitHub issue events, including the type of event, the username that did the event, and the time of the event. If the event is "labeled" or "unlabeled", then the \code{label} is also returned.
}
\examples{
\dontrun{
get_issue_events("https://github.com/ropensci/onboarding/issues/33")
}
}
