% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_events.R
\name{get_raw_events}
\alias{get_raw_events}
\title{Get raw GitHub issue events}
\usage{
get_raw_events(issue_url, limit = Inf)
}
\arguments{
\item{issue_url}{GitHub issue}

\item{limit}{Number of events to return. Defaults to \code{Inf}}
}
\description{
Get a raw list of events associated with a GitHub issue URL.
}
\examples{
\dontrun{
get_raw_events("https://github.com/ropensci/onboarding/issues/33", limit = 1)
}
}
