% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_events.R
\name{parse_events}
\alias{parse_events}
\title{Parse GitHub event}
\usage{
parse_events(event)
}
\arguments{
\item{event}{GitHub event from \code{get_raw_events}}
}
\description{
Parse GitHub event
}
\examples{
\dontrun{
res <- get_raw_events("https://github.com/ropensci/onboarding/issues/33", limit = 2)
parse_events(res[[1]])
}
}
